Move GetIPRange to iphelpers
